Special cased when there are less than three product codes
When there are one or two product codes it is faster to search
for the product code in the line rather than splitting the
line and checking to see if the split line's product code is
in the set of product codes.

There was also a bug in the changefile getting parsed immediately.
Because of how iterdir iterates, this wasn't a problem in practice
but definitely wasn't doing what I wanted.