bug fixing, updating lsp, updating sanity checking
A bunch of off by one errors that I created and never ran into
in the limited testing.  The sanity check for the line lengths
now checks the entire maude data set that was parsed to ensure that
it's catching my stupid mistakes for the special casing.

black and the lsp didn't agree on if a space before a colon should
be allowed, so i let black do its thing.