Updated patient code parsing
The FDA change the file name and format for the patient problem codes.
The new parser is more flexible and allows the FDA to continue to add
columns in the future to hopefully future proof this a little batter.

Also updated the main funtion to return an integer and then
exit base on the returned error.  When instrumenting the code
for profiling this will allow wrapping the main function
and not have to make bigger adjustments due to main exiting on its
own.