Added summarizing data.
This scans through the collected maude data and counts all of the
encountered problems, writing them out to the terminal
as well as dumping to a text file.

Other minor cleanup work like cathing typos and renaming some
variables to make things more sane were also incorporated in this
change.