BUGFIX: Manually merge the paitient problems
A single patient problem shows up per line.

This means that an MDR key can show up multiple times in the
patient problem file.  The chunk parsing handled the merging when
there are multiple problems for a single key, but once the chunks
were returned from the workers there was no checking if an mdr key
shows up in seperate chunks.

This fixes the cases were an mdr key spans multiple chunks.