BUGFIX: fix the custom delimited.
b"  " is used as the delimiter for stiching patient problems together.
When manual merging of chunks was added, this was accidently set to
b" ".  A better seperator should probably be figured out at some point,
but since it's going to get written to a file, using weird characters
will look like shit when someone is analyzing the output data.